split linked portions of the page into their own files
