+++ src/content/apiDocs/oauth/oauthTechnical.mdx
@@ -1,0 +1,38 @@
+---
+imports:
+  '{ CodeWrapper, HeadingWrapper }': ../../../components 
+  '{ HashLink }' : 'react-router-hash-link'
+  '{ onHashAnchorClick }' : '../../../utils/clickHandlers'
+  'GettingStarted': ./oauthTechnicalGettingStarted.mdx
+  'BuildingOpenId': ./oauthTechnicalBuildingOpenId.mdx
+  'Scopes': ./oauthTechnicalScopes.mdx
+  'IdToken': ./oauthTechnicalIdToken.mdx
+  'TestUsers': ./oauthTechnicalTestUsers.mdx
+---
+
+This page walks you through the steps toward authorization and authentication for using OAuth APIs on the Lighthouse platform. We provide examples throughout this document that you can copy and use to get started. 
+
+They are shown with generic cURL commands. Select your API from the drop-down near each cURL command to populate the examples with your API-specific URLs, parameters, and scopes.
+
+## On this Page: {tabIndex="-1"}
+
+- <HashLink to="#getting-started" onClick={onHashAnchorClick}>Getting Started, Initiating the Authorization Code Flow</HashLink>
+- <HashLink to="#building-openid" onClick={onHashAnchorClick}>Building OpenID Connect Applications</HashLink>
+- <HashLink to="#scopes" onClick={onHashAnchorClick}>Scopes</HashLink>
+- <HashLink to="#id-token" onClick={onHashAnchorClick}>ID Token</HashLink>
+- <HashLink to="#test-users" onClick={onHashAnchorClick}>Test Users</HashLink>
+
+### It’s also good to know that: 
+- The access credentials we supply are for the sandbox environment only and will not work in the production environment. 
+- This page provides examples that show authorization server URLs in the sandbox environment, which will differ depending on the API. 
+- When your application is ready, you may [apply for production access](/go-live).
+
+<GettingStarted />
+
+<BuildingOpenId />
+
+<Scopes />
+
+<IdToken />
+
+<TestUsers />
