+++ src/content/apiDocs/oauth/oauthTechnicalIdToken.mdx
@@ -1,0 +1,27 @@
+---
+imports:
+  '{ HeadingWrapper }': ../../../components 
+---
+
+<HeadingWrapper heading="ID Token" id="id-token"/>
+
+Access tokens and id_tokens are [JSON Web Tokens](https://jwt.io/) or JWTs. A JWT consists of three parts: a header, a payload, and a signature. 
+
+An id_token is available in the authorization code flow (response_type = code) token response when the openid scope is used.
+
+Your application must validate JWT signatures. This allows your application to verify that the provided JWT originates from our authorization servers and prevents your application from accepting a JWT with claims that are attempting to impersonate one of your users.
+
+### Header
+
+The JWT's header has two fields, alg and kid. alg indicates the algorithm that was used the sign the JWT, and kid identifies the key that was used to sign the JWT. Signing keys and associated metadata are accessible from 
+https://sandbox-api.va.gov/oauth2/.well-known/openid-configuration.
+
+### Signature
+
+The signature is a cryptographically generated signature of the JWT's header and payload used to confirm the JWT's authenticity. Your application must validate this signature using the alg and the kid from the JWT's header. You may want use one of the JWT libraries listed at [jwt.io](https://jwt.io) to help make this process easier.
+
+### Payload
+
+The payload is a JSON object containing identity and authentication-related claims. There are a couple claims in the JWT that are important for your application to consider:
+- nonce - should match the nonce you initiated authorization with.
+- exp - the expiration time of the JWT. The token cannot be accepted by the Lighthouse platform after this time, and your application should not use an expired token to identify a user.
