+++ src/content/apiDocs/oauth/oauthTechnicalLegacy.mdx
@@ -1,6 +1,6 @@
 ---
 imports:
-  '{ CodeWrapper }': ../../components
+  '{ CodeWrapper }': ../../../components
 ---
