+++ src/content/apiDocs/oauth/oauthTechnicalScopes.mdx
@@ -1,0 +1,25 @@
+---
+imports:
+  '{ HeadingWrapper }': ../../../components 
+---
+
+<HeadingWrapper heading="Scopes" id="scopes"/>
+
+Scopes define the API endpoint your application is allowed to access. We suggest requesting the fewest number of scopes for which you require a Veteran to provide consent. You can always request access to additional scopes if a Veteran or VSO needs the data while using your application.  
+
+**Existing Scopes are:**
+
+| Scope                    | Values and description                                                                                                                   |
+| ------------------------ | -----------------------------------------------------------------------------------------------------------------------------------------|
+| profile                  | Granted by default, allows access to a user's first and last name and email.                                                             |
+| offline_access           | This scope causes the authorization server to provide a refresh token when the [access token](#id-token) is requested.                   |
+| openid                   | An id_token is available in the authorization code flow (response_type = code) token response when the 'openid' scope is used.          |
+
+<!-- (Dynamic Portion Below - API selector) -->
+| Scope                    | Values and description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
+| ------------------------ | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| launch/patient           | A permission setting to obtain the patient's identifier in the token response when the app is launched from an EHR.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                |
+| patient/*                | <p>View a user's VA Health records and patient information, see specific read only scopes below.</p><ul><li>patient/Patient.read</li><li>patient/AllergyIntolerance.read</li><li>patient/Condition.read</li><li>patient/CoverageEligibilityResponse.read</li><li>patient/CommunityCareEligibility.read</li><li>patient/DiagnosticReport.read</li><li>patient/Immunization.read</li><li>patient/Medication.read</li><li>patient/MedicationOrder.read</li><li>patient/MedicationStatement.read</li><li>patient/Observation.read</li><li>patient/Procedure.read</li></ul> |
+| service_history.read     | View a Veteran's service history including deployments and discharge status                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
+| disability_rating.read   | View a Veteran's VA disability ratings and the effective date of the rating                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
+| veteran_status.read      | Confirm the Veteran status of an individual
